# perf: Enhance test suite performance and reliability infrastructure