--- a/app/components/auth_modal_component.html.erb
+++ b/app/components/auth_modal_component.html.erb
@@ -30,12 +30,14 @@
         <div>
           <%= form.label :first_name, "First name", class: "block text-sm font-medium text-gray-700 mb-1" %>
           <%= form.text_field :first_name,
+              required: register_mode?,
               class: "w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500",
               placeholder: "First name" %>
         </div>
         <div>
           <%= form.label :last_name, "Last name", class: "block text-sm font-medium text-gray-700 mb-1" %>
           <%= form.text_field :last_name,
+              required: register_mode?,
               class: "w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500",
               placeholder: "Last name" %>
         </div>
@@ -61,7 +63,7 @@
       <div data-auth-field="register" style="<%= 'display: none;' if login_mode? %>">
         <%= form.label :password_confirmation, "Confirm password", class: "block text-sm font-medium text-gray-700 mb-1" %>
         <%= form.password_field :password_confirmation,
-            required: true,
+            required: register_mode?,
             class: "w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500",
             placeholder: "Confirm your password" %>
       </div>