… .github/workflows/ci.yml
… CLAUDE.md
… docs/plans/feature_flags.md
… docs/plans/rails-form-replacement-plan.md
… spec/support/system_test_helpers.rb
… spec/system/authentication_spec.rb
… spec/system/feature_flag_auth_spec.rb